Module gmp_mpfr_sys::C::GMP

This module is a place-holder for documentation; do not use it in code.

GNU MP 6.2.0

Next: Copying, Previous: (dir), Up: (dir) [Index]

GNU MP

This manual describes how to install and use the GNU multiple precision arithmetic library, version 6.2.0.

Copyright 1991, 1993-2016, 2018 Free Software Foundation, Inc.

Permission is granted to copy, distribute and/or modify this document under the terms of the GNU Free Documentation License, Version 1.3 or any later version published by the Free Software Foundation; with no Invariant Sections, with the Front-Cover Texts being “A GNU Manual”, and with the Back-Cover Texts being “You have freedom to copy and modify this GNU Manual, like GNU software”. A copy of the license is included in GNU Free Documentation License.

• Copying	 	GMP Copying Conditions (LGPL).

• Introduction to GMP	 	Brief introduction to GNU MP.

• Installing GMP	 	How to configure and compile the GMP library.

• GMP Basics	 	What every GMP user should know.

• Reporting Bugs	 	How to usefully report bugs.

• Integer Functions	 	Functions for arithmetic on signed integers.

• Rational Number Functions	 	Functions for arithmetic on rational numbers.

• Floating-point Functions	 	Functions for arithmetic on floats.

• Low-level Functions	 	Fast functions for natural numbers.

• Random Number Functions	 	Functions for generating random numbers.

• Formatted Output	 	printf style output.

• Formatted Input	 	scanf style input.

• C++ Class Interface	 	Class wrappers around GMP types.

• Custom Allocation	 	How to customize the internal allocation.

• Language Bindings	 	Using GMP from other languages.

• Algorithms	 	What happens behind the scenes.

• Internals	 	How values are represented behind the scenes.

• Contributors	 	Who brings you this library?

• References	 	Some useful papers and books to read.

• GNU Free Documentation License

• Concept Index

• Function Index

Next: Copying, Previous: (dir), Up: (dir) [Index]

Constants

Algorithms
This constant is a place-holder for documentation; do not use it in code.

CPP_Class_Interface
This constant is a place-holder for documentation; do not use it in code.

Concept_Index
This constant is a place-holder for documentation; do not use it in code.

Contributors
This constant is a place-holder for documentation; do not use it in code.

Copying
This constant is a place-holder for documentation; do not use it in code.

Custom_Allocation
This constant is a place-holder for documentation; do not use it in code.

Floating_point_Functions
This constant is a place-holder for documentation; do not use it in code.

Formatted_Input
This constant is a place-holder for documentation; do not use it in code.

Formatted_Output
This constant is a place-holder for documentation; do not use it in code.

Function_Index
This constant is a place-holder for documentation; do not use it in code.

GMP_Basics
This constant is a place-holder for documentation; do not use it in code.

GNU_Free_Documentation_License
This constant is a place-holder for documentation; do not use it in code.

Installing_GMP
This constant is a place-holder for documentation; do not use it in code.

Integer_Functions
This constant is a place-holder for documentation; do not use it in code.

Internals
This constant is a place-holder for documentation; do not use it in code.

Introduction_to_GMP
This constant is a place-holder for documentation; do not use it in code.

Language_Bindings
This constant is a place-holder for documentation; do not use it in code.

Low_level_Functions
This constant is a place-holder for documentation; do not use it in code.

Random_Number_Functions
This constant is a place-holder for documentation; do not use it in code.

Rational_Number_Functions
This constant is a place-holder for documentation; do not use it in code.

References
This constant is a place-holder for documentation; do not use it in code.

Reporting_Bugs
This constant is a place-holder for documentation; do not use it in code.