Module gmp_mpfr_sys::C::MPFR

This module is a place-holder for documentation; do not use it in code.

GNU MPFR 4.1.0

Next: Copying, Previous: (dir), Up: (dir) [Index]

GNU MPFR

This manual documents how to install and use the Multiple Precision Floating-Point Reliable Library, version 4.1.0.

Copyright 1991, 1993-2020 Free Software Foundation, Inc.

Permission is granted to copy, distribute and/or modify this document under the terms of the GNU Free Documentation License, Version 1.2 or any later version published by the Free Software Foundation; with no Invariant Sections, with no Front-Cover Texts, and with no Back-Cover Texts. A copy of the license is included in GNU Free Documentation License.

• Copying	 	MPFR Copying Conditions (LGPL).

• Introduction to MPFR	 	Brief introduction to GNU MPFR.

• Installing MPFR	 	How to configure and compile the MPFR library.

• Reporting Bugs	 	How to usefully report bugs.

• MPFR Basics	 	What every MPFR user should now.

• MPFR Interface	 	MPFR functions and macros.

• API Compatibility	 	API compatibility with previous MPFR versions.

• MPFR and the IEEE 754 Standard

• Contributors

• References

• GNU Free Documentation License

• Concept Index

• Function and Type Index

Next: Copying, Previous: (dir), Up: (dir) [Index]

Constants

API_Compatibility
This constant is a place-holder for documentation; do not use it in code.

Concept_Index
This constant is a place-holder for documentation; do not use it in code.

Contributors
This constant is a place-holder for documentation; do not use it in code.

Copying
This constant is a place-holder for documentation; do not use it in code.

Function_and_Type_Index
This constant is a place-holder for documentation; do not use it in code.

GNU_Free_Documentation_License
This constant is a place-holder for documentation; do not use it in code.

Installing_MPFR
This constant is a place-holder for documentation; do not use it in code.

Introduction_to_MPFR
This constant is a place-holder for documentation; do not use it in code.

MPFR_Basics
This constant is a place-holder for documentation; do not use it in code.

MPFR_Interface
This constant is a place-holder for documentation; do not use it in code.

MPFR_and_the_IEEE_754_Standard
This constant is a place-holder for documentation; do not use it in code.

References
This constant is a place-holder for documentation; do not use it in code.

Reporting_Bugs
This constant is a place-holder for documentation; do not use it in code.