Struct gmp_mpfr_sys::gmp::mpf_t

#[repr(C)]pub struct mpf_t {
    pub prec: c_int,
    pub size: c_int,
    pub exp: exp_t,
    pub d: NonNull<limb_t>,
}

See: mpf_t and Float Internals

Future compatibility

The fields listed here are considered internal details. These internals may change in new minor releases of this crate, though they will be kept unchanged for patch releases. Any code that makes use of these internals should list the dependency as version = "~1.3" inside Cargo.toml, not version = "1.3".

Fields
prec: c_int

See: Float Internals

size: c_int

See: Float Internals

exp: exp_t

See: Float Internals

d: NonNull<limb_t>

See: Float Internals