gmp_mpfr_sys::mpc

Struct mpcb_t

#[repr(C)]
pub struct mpcb_t {
    pub c: mpc_t,
    pub r: mpcr_t,
}

Experimental struct. See: mpcb_t

Future compatibility

The fields listed here are considered internal details. These internals may change in new minor releases of this crate, though they will be kept unchanged for patch releases. Any code that makes use of these internals should list the dependency as version = "~1.5" inside Cargo.toml, not version = "1.5".

Fields
c: mpc_t

Internal implementation detail: center.

r: mpcr_t

Internal implementation detail: radius.