Enum gmp_mpfr_sys::mpfr::rnd_t

#[repr(C)]
pub enum rnd_t {
    RNDN,
    RNDZ,
    RNDU,
    RNDD,
    RNDA,
    RNDF,
    // some variants omitted
}

See: mpfr_rnd_t

Planned change

In the next major version of the crate (version 2), this enum will be replaced by a type alias to c_int. The variants will be replaced by constants, for example rnd_t::RNDN will be replaced by a constant RNDN of type c_int.

Variants

RNDN

See: Rounding

Planned change

In the next major version of the crate (version 2), the enum will be removed. This variant will be replaced by a constant RNDN of type c_int.

RNDZ

See: Rounding Modes

Planned change

In the next major version of the crate (version 2), the enum will be removed. This variant will be replaced by a constant RNDZ of type c_int.

RNDU

See: Rounding Modes

Planned change

In the next major version of the crate (version 2), the enum will be removed. This variant will be replaced by a constant RNDU of type c_int.

RNDD

See: Rounding Modes

Planned change

In the next major version of the crate (version 2), the enum will be removed. This variant will be replaced by a constant RNDD of type c_int.

RNDA

See: Rounding Modes

Planned change

In the next major version of the crate (version 2), the enum will be removed. This variant will be replaced by a constant RNDA of type c_int.

RNDF

See: Rounding Modes

Planned change

In the next major version of the crate (version 2), the enum will be removed. This variant will be replaced by a constant RNDF of type c_int.